Generator done
Generator is now done. Script generates data for all tables in database.

One thing that can be change is the complex save_to_csv function.
It is messy becouse of the previous assumptions that methods generating
data would need to be be generator objects.

When it was done, and the moment of synchronization needed to create
3 tables with the same Id came, the approach happen to be wrong.

Otherway, the script is working fine and is ready to be used.